# Optimization of validator exit processing