State transition voluntary exit timings:

With this PR:
```
 ** exits: 0 ms: 0
 ** exits: 16 ms: 97
 ** exits: 0 ms: 0
 ** exits: 16 ms: 71
 ** exits: 16 ms: 23
 ** exits: 16 ms: 24
```

Previous implementation:
```
 ** exits: 0 ms: 0
 ** exits: 16 ms: 1571
 ** exits: 0 ms: 0
 ** exits: 16 ms: 1399
 ** exits: 16 ms: 1442
 ** exits: 16 ms: 1325
```

fixes: #7041

## Documentation

- [ ] I thought about documentation and added the `doc-change-required` label to this PR if updates are required.

## Changelog

- [ ] I thought about adding a changelog entry, and added one if I deemed necessary.