@tbenr 2023-04-16
Btw there is also the slashing codepath going there..