@tbenr 2023-04-16
actually i'm not super sure about putting BeaconState things in the context.. seems like a bit less clean...