@tbenr 2023-04-16
I'm also thinking that this approach could be generalized for other operations (mutators) if thay could benefit in the same way: image at the beginning of a state transition we lazily create these contexts that are available for each state mutators. But the assumption that there are no interactions, a given state field can be changed by only one mutator so they do not interfere, so contexts contain completely different set of values.

Just an attempt to generalize the approach,if make sense.