@rolfyone 2023-04-16
I guess the real question would be if this improved numbers further? :)