@tbenr 2023-04-17
A different approach to this would be to do something similar to `TransitionCaches` on the state.